README Install Instructions Correction
In the install directions here in the README, it says to install eslint-config-semistandard twice. 

I'm pretty sure the first line was supposed to be eslint-config-standard since it gets marked as an UNMET PEER DEPENDENCY by npm 3.

There are also 2 unmet peer dependencies for the standard config (its plugin and the promise plugin), which I included with the eslint-plugin-react you already had listed.

Additionally, the specific version eslint-config-standard@5.1.0 was being asked for, so I tacked that on. I think you will be fixing that soon though based on what I read in the other open issue.

Thanks,
-Josh